requirements.txt
pygame>=2.5.0
mutagen>=1.47.0
flask>=2.3.0
flask-socketio>=5.3.0
Pillow>=10.0.0
requests>=2.31.0
psutil>=5.9.0